test/web-platform-tests/to-upstream/css/selectors/i18n/css3-selectors-lang-041-dont-upstream.html

<!DOCTYPE html>
<!-- jsdom only test without offsetWidth -->
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>[lang="es"], lang="es"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#attribute-selectors'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 50px; }
#box[lang='es'] { width: 100px; }
</style>
</head>
<body>

<div class="test"><div id="box" lang="es">&#xA0;</div></div>

<script>
"use strict";
test(() => {
  assert_true(document.getElementById("box").matches("#box[lang='es']"));
}, "A lang= value that matches an identical lang attribute value on the same element will produce styling.");
</script>

<div id='log'></div>

</body>
</html>
